requirements.txt
langchain-groq
langchain_core
python-dotenv
streamlit
langchain_community
langserve
langchain-ollama